remove enemy objects from game after leaving screen
In the script.js file, added functionality to remove enemy objects from the game after they move off the screen. This was implemented by adding a 'markedForDeletion' property to the Enemy class, which is set to true when the enemy's x coordinate is less than 0 minus its width. The game's update method now filters out any enemies marked for deletion. This should improve performance by reducing the number of objects in memory.